fix long running transform bug (does not affect scaling)
The transforms need to be built in reverse. Intuition here is that you
have a point in its own coordinate system, and it must go through a series
of transforms to get to the overall SVG's coordinate system. That goes
from its last transform to the first transform of the svg.